+++ providers/merry/provider.go
@@ -9,9 +9,9 @@ package merry
 import (
 	"github.com/ansel1/merry"
 
-	"github.com/myhelix/contextlogger/log"
-	"github.com/myhelix/contextlogger/providers"
-	"github.com/myhelix/contextlogger/providers/chaining"
+	"github.com/calm/contextlogger/log"
+	"github.com/calm/contextlogger/providers"
+	"github.com/calm/contextlogger/providers/chaining"
 
 	"context"
 )