+++ providers/rollbar/provider.go
@@ -9,12 +9,12 @@ is a merry error; otherwise it generates a new one based on the reporting callst
 package rollbar
 
 import (
+	"github.com/calm/rollbar"
 	goerr "github.com/go-errors/errors"
-	"github.com/myhelix/rollbar"
 
-	"github.com/myhelix/contextlogger/log"
-	"github.com/myhelix/contextlogger/providers"
-	"github.com/myhelix/contextlogger/providers/chaining"
+	"github.com/calm/contextlogger/log"
+	"github.com/calm/contextlogger/providers"
+	"github.com/calm/contextlogger/providers/chaining"
 
 	"context"
 	"errors"